HW2 might be complete, run more tests
