+++ app/src/main/java/com/example/robotapp20/MainActivity.kt
@@ -18,14 +18,8 @@ class MainActivity : AppCompatActivity() {
     private lateinit var whiteRobot: ImageView
 
     private lateinit var robotImages: MutableList<ImageView>
-
-    private val robots = listOf(
-        Robot(false, R.drawable.robot_red_large, R.drawable.robot_red_small),
-        Robot(false, R.drawable.robot_white_large, R.drawable.robot_white_small),
-        Robot(false, R.drawable.robot_yellow_large, R.drawable.robot_yellow_small)
-    )
-
     private val robotViewModel: RobotViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -33,50 +27,40 @@ class MainActivity : AppCompatActivity() {
         yellowRobot = findViewById(R.id.yellow_robot_large)
         redRobot = findViewById(R.id.red_robot_large)
         whiteRobot = findViewById(R.id.white_robot_large)
-
         robotImages = mutableListOf(redRobot, whiteRobot, yellowRobot)
 
+        //sets images if app is re-opened or rotated
+        if (robotViewModel.turnCount != 0){
+            setImages()
+        }
+
         yellowRobot.setOnClickListener {
-            advanceTurn()
+            robotViewModel.advanceTurn()
             Toast.makeText(this, "TurnCount : ${robotViewModel.turnCount}", Toast.LENGTH_SHORT)
                 .show()
+            setImages()
         }
 
         redRobot.setOnClickListener {
-            advanceTurn()
+            robotViewModel.advanceTurn()
+            setImages()
         }
 
         whiteRobot.setOnClickListener {
-            advanceTurn()
+            robotViewModel.advanceTurn()
+            setImages()
         }
 
     }// end of onCreate
 
-
-    private fun advanceTurn() {
-        robotViewModel.turnCount += 1
-        if (robotViewModel.turnCount > 3) {
-            robotViewModel.turnCount = 1
-        }
-        setRobotTurn()
-        setImages()
-    }
-
     private fun setImages() {
-        for (indy in robots.indices) {
-            if (robots[indy].myTurn) {
-                robotImages[indy].setImageResource(robots[indy].largeImgRes)
+        for (indy in robotViewModel.robots.indices) {
+            if (robotViewModel.robots[indy].myTurn) {
+                robotImages[indy].setImageResource(robotViewModel.robots[indy].largeImgRes)
             } else {
-                robotImages[indy].setImageResource(robots[indy].smallImgRes)
+                robotImages[indy].setImageResource(robotViewModel.robots[indy].smallImgRes)
             }
         }
     }
 
-
-    private fun setRobotTurn() {
-        for (robot in robots) {
-            robot.myTurn = false
-        }
-        robots[robotViewModel.turnCount - 1].myTurn = true
-    }
 }
